code cleanup and javadoc update of MSK-464 code
