+++ moskito-core/src/main/java/net/anotheria/moskito/core/producers/ICustomDecoratorStats.java
@@ -2,9 +2,25 @@ package net.anotheria.moskito.core.producers;
 
 import net.anotheria.moskito.core.decorators.IDecoratorFactory;
 
-public interface ICustomDecoratorStats {
+/**
+ * Interface for stats that requires
+ * object-specific decorators instead
+ * of class-specific.
+ */
+public interface ICustomDecoratorStats<SELF extends ICustomDecoratorStats<SELF>> extends IStats {
 
+    /**
+     * Returns id of decorator that can be used with
+     * this stats object
+     * @return id of decorator
+     */
     String getDecoratorId();
-    IDecoratorFactory getDecoratorFactory();
+
+    /**
+     * Returns decorator factory to build new
+     * decorator if it not present in registry yet.
+     * @return instance of decorator factory for this stats object
+     */
+    IDecoratorFactory<SELF> getDecoratorFactory();
 
 }
